Remove old client, fix MockClient
